refactoring files
- create MVC like file structure
- move files to it corresponding folder
- update file "requires"
