+++ common/corrector.rb
