+++ main.rb
@@ -1,4 +1,3 @@
-require 'pry'
 require './uiapp'
 
 def main
