+++ model/book.rb
