+++ model/rental.rb
