+++ model/student.rb
@@ -1,5 +1,5 @@
-require './person'
-require './classroom'
+require_relative './person'
+require_relative './classroom'
 
 class Student < Person
   def initialize(age, name = 'unknown', parent_permission: true)
